Promote RetryStrategy to shared; jitter the WS reconnect
Unifies the connection-retry policy (Tier 2 #5, retry-policy half). The
applet's RetryStrategy (exponential + ±10% jitter — the best of the three
policies) moves to super-stt-shared::daemon::retry. The applet now uses it
directly (its local retry.rs + test deleted), and the shared
widget-subscription reconnect loop drives it via next_delay/should_retry/
reset instead of the jitter-less next_backoff doubling — so the SSE
reconnect gains jitter too, preventing a thundering herd of clients after a
daemon restart. Also guards a latent `% 0` panic in next_delay when the
jitter window rounds to zero (sub-10ms initial delays).

Remaining for #5: the settings app still reconnects on a flat 5s sleep
(adopting RetryStrategy there needs an attempt-counter field on AppModel),
and the per-crate subscription bridges are unchanged.
