+++ super-stt-cosmic-applet/src/daemon/mod.rs
@@ -1,8 +1,5 @@
 // SPDX-License-Identifier: GPL-3.0-only
 pub mod client;
-pub mod retry;
-#[cfg(test)]
-mod retry_test;
 
 pub use client::*;
-pub use retry::RetryStrategy;
+pub use super_stt_shared::daemon::retry::RetryStrategy;
