+++ super-stt-shared/src/daemon/mod.rs
@@ -2,5 +2,6 @@
 //! Shared daemon communication functionality for Super STT applications
 
 pub mod http_client;
+pub mod retry;
 pub mod session;
 pub mod widget_subscription;
